Small Improvements

+++ Backend/src/index.ts
@@ -7,11 +7,13 @@ type Task = {
   desc: string;
 };
 
-type TaskTodo = {
-  id: string;
-  name: string;
-  desc: string;
-};
+function GenerateId(): string {
+  let id: string;
+  do {
+    id = Math.floor(10000 + Math.random() * 90000).toString();
+  } while (DB.has(id));
+  return id;
+}
 
 const DB = new Map<string, Task>();
 
@@ -21,9 +23,11 @@ app.post("/task", async (c) => {
     const body = await c.req.json();
     const { name, desc } = body;
 
-    if (!name || !desc) return c.json({ message: "Invalid Data" }, 400);
+    if (!name || !desc) {
+      return c.json({ message: "Invalid Data" }, 400);
+    }
 
-    const id = Math.floor(Math.random() * 1000).toString();
+    const id = GenerateId();
 
     DB.set(id, { name, desc });
 
@@ -39,15 +43,12 @@ app.post("/task", async (c) => {
 
 app.get("/task", async (c) => {
   try {
-    const tasks: TaskTodo[] = [];
+    const tasks = Array.from(DB, ([id, task]) => ({
+      id,
+      name: task.name,
+      desc: task.desc,
+    }));
 
-    DB.forEach((value, key) => {
-      tasks.push({
-        id: key,
-        name: value.name,
-        desc: value.desc,
-      });
-    });
     return c.json({
       message: "get method Created successfully",
       data: tasks,
@@ -61,7 +62,7 @@ app.get("/task", async (c) => {
 app.get("/task/:id", async (c) => {
   try {
     const { id } = c.req.param();
-    
+
     if (!DB.has(id)) {
       return c.json({ message: "Id Not Found" }, 400);
     }
@@ -69,17 +70,20 @@ app.get("/task/:id", async (c) => {
     const task = DB.get(id);
     return c.json({
       message: "Get Method By Id working Successfully",
-      data: { id, task}
+      data: { id, task },
     });
   } catch (error) {
     console.log(error);
     return c.json({ message: "Internal Server Error" }, 500);
   }
-})
+});
 
-app.put("/task/:id", async (c) => {
+app.put("/task", async (c) => {
   try {
-    const { id } = c.req.param();
+    const id = c.req.query("id");
+    if (!id) {
+      return c.json({ message: "Id Not Found" }, 400);
+    }
 
     if (!DB.has(id)) {
       return c.json({ message: "Id Not Found" }, 400);
@@ -102,18 +106,23 @@ app.put("/task/:id", async (c) => {
   }
 });
 
-app.delete("/task/:id", async (c) => {
+app.delete("/task", async (c) => {
   try {
-    const { id } = c.req.param();
+    const id = c.req.query("id");
+
+    if (!id) {
+      return c.json({ message: "Id Not Found" }, 400);
+    }
 
     if (!DB.has(id)) {
       return c.json({ message: "Id Not Found" }, 400);
     }
 
     DB.delete(id);
+
     return c.json({
-      message: "delete method Working successfully"
-    })
+      message: "delete method Working successfully",
+    });
   } catch (error) {
     console.log(error);
     return c.json({ message: "Internal Server Error" }, 500);
